refactor(code-editor): reorganize test utilities and update imports
