+++ packages/code-editor/src/index.ts
@@ -1,9 +1,3 @@
-import { renderCodeEditor } from './CodeEditor';
-
-export const TestUtils = {
-  renderCodeEditor,
-};
-
 export {
   CodeEditor,
   type CodeEditorProps,
